==> 02-univariate.Rmd <==
# Univariate Descriptives {#Uni.Descriptives}
```{r desc.setup, include=FALSE}
library(bookdown)
library(tidyverse)
library(knitr)
library(gridExtra)
library(kableExtra)
library(table1)
library(gridExtra)
```

```{r data.prep, include = F}
# Preparing colours for graphs
.grey <- "#A6A6A6"
.ma <- "#DC2D27"
.nma <- "#39B9FF"
.lav <- "#D9BBDC"
.lgreen <- "#BEDCB3"

# Data
load("data/objects/data.full.Rda")
load("data/objects/screened.full.Rda")
load("data/objects/screened.Rda")

# Changing NA from meth_ingest and meth_user to "No"
screened$meth_ingest[is.na(screened$meth_ingest)] <- "No"
screened$meth_user[is.na(screened$meth_user)] <- "Non-User"
screened$meth_user <- as.factor(screened$meth_user)

# Rearranging levels for education 
screened$education <- factor(screened$education, levels = c("Did not finish high school",
                                                          "High School Diploma",
                                                          "Vocational/Technical degree or certificate",
                                                          "Did not finish University",
                                                          "Bachelor Degree",
                                                          "Postgraduate Degree"))
```
This section includes Univariate Descriptives (stratified by Methamphetamine Use) for the key variables in the study.

***

## Demographic Summary Table
```{r desc.summary, echo = F}
x <- screened %>% mutate(ethnicity = reorder(ethnicity, ethnicity, function(x) -length(x)),
                        meth_user = reorder(meth_user, meth_user, function(x) length(x)),
                        area_live = reorder(area_live, area_live, function(x) length(x)))
label(x$age) <- "Age"
label(x$sex) <- "Sex"
label(x$meth_ingest) <- "MA Ingest"
label(x$meth_user) <- "MA User Type"
label(x$license_status) <- "License Status"
label(x$education) <- "Education"
label(x$ethnicity) <- "Ethnicity"
label(x$area_live) <- "Area of Residence"


table1(~age + sex + meth_ingest + meth_user + license_status + education + ethnicity + area_live, data = x) # %>% 
  # scroll_box(height = "500px")

```

***

## Distribution of Demographics 

### Sex
```{r distr.sex, echo = F}
bar.sex <- screened %>% 
  ggplot(aes(x = sex, fill = sex)) +
  geom_bar(col = "black") +
  geom_text(stat = "count", aes(label = ..count..), vjust = -.3) +
  labs(x = "Sex", y = "") +
  ggtitle("Sex Count") +
  theme_light() +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.lav, .lgreen))

bar.sex.ma <- screened %>%
  ggplot(aes(x = sex, fill = meth_ingest)) +
  geom_bar(col = "black") +
  geom_text(stat = "count", aes(label = ..count..), vjust = -.3) +
  labs(x = "Sex", y = "n") +
  ggtitle("Sex count by MA use") +
  theme_light() +
  theme(legend.position = "none") +
  facet_wrap(~ meth_ingest) +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(bar.sex, bar.sex.ma, nrow = 1)

```

***

### Age
```{r distr.age, echo = F}
# Histogram age
hist.age <- screened %>% 
  ggplot(aes(x = age)) +
  geom_histogram(binwidth = 2, col = "black", fill = .grey) +
  labs(x = "Age", y = "") +
  #ggtitle("Distribution of Age") +
  theme_light()

#Histogram of age by MA
hist.age.ma <- screened %>%
  ggplot(aes(x = age, fill = meth_ingest)) +
  geom_histogram(binwidth = 1, col = "black") +
  labs(x = "Age", y = "", fill = "MA Use") +
  #ggtitle("Distribution of Age by MA use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  scale_fill_manual(values = c(.ma, .nma))

# Density plot of age by MA
dens.age.ma <- screened %>% 
  ggplot(aes(x = age, fill = meth_ingest)) +
  geom_density(alpha = 0.5) +
  labs(x = "Age", y = "Density", fill = "MA Use") +
  #ggtitle("Distribution of Age by MA use (Density)") +
  theme_light() +
  scale_fill_manual(values = c(.ma, .nma))

# Boxplot of age by MA
box.age.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y =  age, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  scale_fill_manual(values = c(.ma, .nma)) +
  geom_jitter(alpha = 0.5, width = 0.08) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Age");
  #ggtitle("Distribution of Age by MA use")

grid.arrange(hist.age, hist.age.ma, dens.age.ma, box.age.ma, top = "Distribution of Age by MA Use")
```

***

### Education
```{r distr.edu, echo = F}
bar.edu <- screened %>% 
  mutate(education = fct_reorder(education, desc(education))) %>% 
  ggplot(aes(x = education)) +
  geom_bar(col = "black", fill = .grey) +
  coord_flip() +
  labs(x = "Education Level", y = "") +
  #ggtitle("Education Count") +
  theme_light()

bar.edu.ma <- screened %>%
  mutate(education = fct_reorder(education, desc(education))) %>% # Reorder levels of education 
  ggplot(aes(x = education, fill = meth_ingest)) +
  geom_bar(position = "dodge", col = "black") +
  #geom_text(stat = "count", aes(label = ..count..)) +
  coord_flip() +
  labs(x = "Education Level", y = "", fill = "MA Use") +
  #ggtitle("Education count by MA use") +
  theme_light() +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(bar.edu, bar.edu.ma, nrow = 2, top = "Education Distribution")
```

***

### Ethnicity
```{r distr.eth, echo = F}
bar.eth <- screened %>% 
  ggplot(aes(x = reorder(ethnicity, ethnicity, function(x) -length(x)))) +
  geom_bar(fill = .grey, col = "black") +
  geom_text(stat = "count", aes(label = ..count..), hjust = -.3) +
  coord_flip() +
  labs(x = "Ethnicity", y = "") +
  theme_light()

bar.eth.ma <- screened %>%
  ggplot(aes(x = reorder(ethnicity, ethnicity, function(x) -length(x)), fill = meth_ingest)) +
  geom_bar(col = "black", position = "dodge") +
  coord_flip() +
  labs(x = "Ethnicity", y = "", fill = "MA Use") +
  theme_light() +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(bar.eth, bar.eth.ma, nrow = 2, top = "Ethnicity Distribution")
```

***

### Area of Residence
```{r distr.live, echo = F}
bar.res <- screened %>% 
  ggplot(aes(x = reorder(area_live, area_live, function(x) length(x)))) +
  geom_bar(fill = .grey, col = "black") +
  geom_text(stat = "count", aes(label = ..count..), vjust = -.3) +
  labs(x = "Area of Residence", y = "") +
  theme_light()

bar.res.ma <- screened %>% 
  ggplot(aes(x = reorder(area_live, area_live, function(x) length(x)), fill = meth_ingest)) +
  geom_bar(col = "black", position = "dodge") +
  labs(x = "Area of Residence", fill = "MA Use") +
  scale_fill_manual(values = c(.ma, .nma)) +
  theme_light();

grid.arrange(bar.res, bar.res.ma, nrow = 2, top = "Area of Residence Distribution")

```

***
## Drug Use
### Alcohol
```{r, echo=FALSE}
# Rearranging levels
levels(screened$AUDIT_frequency) <- c("Four or more times a week", "Two to three times per week", "Two to four times a month","Monthly or less", "Never")

label(screened$alcohol_ever) <- "Have you ever consumed Alcohol?"
label(screened$AUDIT_frequency) <- "How often (if at all) do you consume Alcohol"

table1(~alcohol_ever, data = screened, caption = "Alcohol Consumption (Overall)")
table1(~alcohol_ever, data = screened %>% filter(meth_ingest == "Yes"), caption = "Alcohol Consumption (MA Users)")
table1(~alcohol_ever, data = screened %>% filter(meth_ingest == "No"), caption = "Alcohol Consumption (non-drug users)")

# Frequency (covered in AUDIT)
# table1(~AUDIT_frequency, data = screened, caption = "Frequency of Alcohol Use (Total)")
# table1(~AUDIT_frequency, data = screened %>% filter(meth_ingest == "Yes"), caption = "Frequency of Alcohol Use (MA Users)")
# table1(~AUDIT_frequency, data = screened %>% filter(meth_ingest == "No"), caption = "Frequency of Alcohol Use (Non-Drug Users)")
# 
# alc.freq.ma <- screened %>% 
#   filter(meth_ingest == "Yes") %>% 
#   ggplot(aes(x = AUDIT_frequency, fill = meth_ingest)) +
#   geom_bar(aes(y = (..count..)/sum(..count..)*100)) +
#   coord_flip() +
#   scale_fill_manual(values = .ma) +
#   theme_light() +
#   theme(legend.position = "none") +
#   labs(x = "", y = "Proportion (%)") +
#   ggtitle("Alcohol Frequency in MA Users")
# 
# alc.freq.nma <- screened %>% 
#   filter(meth_ingest == "No") %>% 
#   ggplot(aes(x = AUDIT_frequency, fill = meth_ingest)) +
#   geom_bar(aes(y = (..count..)/sum(..count..)*100)) +
#   coord_flip() +
#   scale_fill_manual(values = .nma) +
#   theme_light() +
#   theme(legend.position = "none") +
#   labs(x = "", y = "Proportion (%)") +
#   ggtitle("Alcohol Frequency in non-drug users")
# 
# grid.arrange(alc.freq.ma, alc.freq.nma)
```

#### AUDIT-C {-}
The <a href = https://cde.drugabuse.gov/instrument/f229c68a-67ce-9a58-e040-bb89ad432be4> Alcohol Use Disorders Identification Test-Concise (AUDIT-C) </a> is a brief alcohol screening instrument that reliably identifies persons who are hazardous drinkers or have active alcohol use disorders (including alcohol abuse or dependence).

The AUDIT-C has 3 questions and is scored on a scale of 0-12. Each AUDIT-C question has 5 answer choices valued from 0 points to 4 points. In men, a score of 4 or more is considered positive, optimal for identifying hazardous drinking or active alcohol use disorders. In women, a score of 3 or more is considered positive. Generally the higher the score, the more likely it is that a person's drinking is affecting his or her safety.

```{r AUDIT.Prep, include = F}
label(screened$AUDIT_Total) <- "AUDIT-C Total"
label(screened$AUDIT_Risky) <- "AUDIT-C Risky"

# Proportional bar chart function
bar.prop <- function(df, xcol, ycol, fill) {
  ggplot(data = df, aes(x = {{xcol}}, fill = {{fill}})) +
    geom_bar(aes(y = (..count..) / sum(..count..) * 100)) +
    theme_light()
}

bar.audit <- bar.prop(screened %>% filter(!is.na(AUDIT_Risky)), xcol = AUDIT_Risky) +
  scale_fill_manual(values = .grey) +
  labs(x = "AUDIT-C Hazardous Drinking", y = "Proportion (%)") +
  ggtitle("AUDIT-C High Score (>3 Men, >2 Women)")
  
  
bar.audit.ma <- screened %>% 
  filter(meth_ingest == "Yes", !is.na(AUDIT_Risky)) %>% 
  ggplot(aes(x = AUDIT_Risky, fill = meth_ingest)) +
  geom_bar(aes(y = (..count..)/sum(..count..)*100)) +
  scale_fill_manual(values = .ma) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "AUDIT-C Hazardous Drinking", y = "Proportion (%)") +
  ggtitle("AUDIT-C High Score (MA Users)")

bar.audit.nma <- bar.prop(screened %>% filter(meth_ingest == "No", !is.na(AUDIT_Risky)), xcol = AUDIT_Risky, fill = meth_ingest) +
  scale_fill_manual(values = .nma) +
  theme(legend.position = "none") +
  labs(x = "AUDIT-C Hazardous Drinking", y = "Proportion (%)") +
  ggtitle("AUDIT-C High Score (Non-Drug Users)")
```

##### Distribution of AUDIT-C Scores {-}
```{r AUDIT.distr, echo = F, warning=F}
hist.audit.all <- screened %>% 
  ggplot(aes(x = AUDIT_Total)) +
  geom_histogram(binwidth = 1, col = "black") +
  theme_light() +
  labs(x = "AUDIT-C Total Score", y = "Count")

dens.audit <- screened %>% 
  ggplot(aes(x = AUDIT_Total, fill = meth_ingest)) +
  geom_density(alpha = 0.5) +
  theme_light() +
  labs(x = "AUDIT-C Total", y = "Density", fill = "MA Use") +
  scale_fill_manual(values = c(.ma, .nma))

hist.audit <- screened %>% 
  ggplot(aes(x = AUDIT_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 1, col = "black") +
  theme_light() +
  facet_grid(meth_ingest ~ .) +
  labs(x = "AUDIT-C Total Score", y = "Count", fill = "Meth Use")

box.audit <- screened %>% 
  ggplot(aes(x = meth_ingest, y = AUDIT_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.5) +
  geom_jitter(height = 0, width = 0.03, alpha = 0.5) +
  theme_light() +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "MA Use", y = "AUDIT-C Total") +
  theme(legend.position = "none")

grid.arrange(hist.audit.all, hist.audit, dens.audit, box.audit, ncol = 2)
```

##### Proportion considered "Hazardous Drinkers" {-}
```{r AUDIT, echo = F}
table1(~AUDIT_Total + AUDIT_Risky, data = screened %>% filter(!is.na(AUDIT_Total)), caption = "AUDIT-C Overall Scores")

table1(~AUDIT_Total + AUDIT_Risky, data = screened %>% filter(meth_ingest == "Yes", !is.na(AUDIT_Total)), caption = "AUDIT-C Scores (MA Users)")

table1(~AUDIT_Total + AUDIT_Risky, data = screened %>% filter(meth_ingest == "No", !is.na(AUDIT_Total)), caption = "AUDIT-C Scores (Non-Drug Users)")

bar.audit
grid.arrange(bar.audit.ma, bar.audit.nma, ncol = 2)
```

### Drugs (MA Users only)

#### Methamphetamine {-}

```{r distr.ma, echo = F}
plot.count <- screened %>%
  ggplot(aes(x = reorder(meth_user, meth_user, function(x) -length(x)))) +
  geom_bar(fill = .grey, col = "black") +
  geom_text(stat = "count", aes(label = ..count..), vjust = -.3) +
  labs(x = "Type of User", y = "n") +
  ggtitle("Count of MA and non-MA users") +
  theme_light()

plot.count
```

```{r ma.prep, include = FALSE}
screened <- screened %>% 
  mutate(meth_use_extreme = fct_recode(as_factor(meth_use_ways),
                                    "Injection" = "Injection under the skin?,Snorting, sniffing, breathing?,By mouth, drinking or in a pill?",
                                    "Injection" = "Injection under the skin?,By mouth, drinking or in a pill?",
                                    "Injection" = "Injection under the skin?,Smoking, freebasing?",
                                    "Injection" = "Injection under the skin?,Smoking, freebasing?,By mouth, drinking or in a pill?",
                                    "Injection" = "IV, through the veins?,Injection under the skin?,By mouth, drinking or in a pill?",
                                    "Injection" = "Injection under the skin?,Smoking, freebasing?,Snorting, sniffing, breathing?,By mouth, drinking or in a pill?",
                                    "Injection" = "IV, through the veins?,Injection under the skin?,Smoking, freebasing?,Snorting, sniffing, breathing?,By mouth, drinking or in a pill?",
                                    "Injection" = "Injection under the skin?",
                                    "Injection" = "IV, through the veins?,By mouth, drinking or in a pill?",
                                    "Injection" = "IV, through the veins?",
                                    "Injection" = "IV, through the veins?,Injection under the skin?",
                                    "Injection" = "IV, through the veins?,Injection under the skin?,Snorting, sniffing, breathing?,By mouth, drinking or in a pill?",
                                    "Injection" = "IV, through the veins?,Smoking, freebasing?,Snorting, sniffing, breathing?,By mouth, drinking or in a pill?",
                                    "Injection" = "IV, through the veins?,Injection under the skin?,Snorting, sniffing, breathing?",
                                    "Smoking" = "Smoking, freebasing?,Snorting, sniffing, breathing?,By mouth, drinking or in a pill?",
                                    "Smoking" = "Smoking, freebasing?,Snorting, sniffing, breathing?",
                                    "Smoking" = "Smoking, freebasing?,By mouth, drinking or in a pill?",
                                    "Smoking" = "Smoking, freebasing?",
                                    "Snorting" = "Snorting, sniffing, breathing?",
                                    "Oral" = "By mouth, drinking or in a pill?"))
         
```

```{r ma.rearrange, include = FALSE}
# Rearranging levels
screened$meth_use_at_peak <- factor(screened$meth_use_at_peak, levels = c("Daily", "Weekly", "1 to 2 times per month", "Less than or about 1 time per month"))
screened$meth_recent_use <- factor(screened$meth_recent_use, levels = c("Days", "Weeks", "Months", "Years"))
screened$meth_use_extreme <- factor(screened$meth_use_extreme, levels = c("Oral", "Snorting", "Smoking", "Injection"))
```

```{r ma.summ, echo = FALSE}
ma.df <- screened %>% filter(meth_ingest == "Yes") # df with only ma users

ma.df <- ma.df %>% 
  mutate(polydrug_use = ifelse(ma.df$drug_use_cannabis != "Never" | 
                                 ma.df$drug_use_club_drugs != "Never" |
                                 ma.df$drug_use_cocaine != "Never" |
                                 ma.df$drug_use_hallucinogens != "Never" |
                                 ma.df$drug_use_heroin != "Never" |
                                 ma.df$drug_use_sedatives != "Never" , "Yes", "No"))

# Labelling for table
label(ma.df$meth_use_at_peak) <- "At the time when you were using MA the most, how often did you use it?"
label(ma.df$meth_last_12_months) <- "In the last 12 months, did you use MA in any form?"
label(ma.df$meth_use_age) <- "How old were you when you first began using MA?"
label(ma.df$meth_recent_use) <- "When was the most recent time you used MA?"
label(ma.df$meth_use_extreme) <- "Most extreme form of use"
label(ma.df$polydrug_use) <- "Ever used an illicit drug other than MA?"

table1(~meth_use_age + meth_use_at_peak + meth_last_12_months + meth_recent_use + meth_use_extreme + polydrug_use , data = ma.df)# %>% 
  # kable_styling("striped") %>% 
  # scroll_box(height = "800px")
```

```{r ma.plots, echo = F}
ma.df %>% 
  ggplot(aes(x = meth_use_age)) +
  geom_histogram(binwidth = 1, col = "black") +
  theme_light() +
  labs(x = "How Old were you when you first began using MA?", y = "")
```

#### Other Drug Use {-}
TBC

***

## Driving Under the Influence
### Alcohol (DUI)
Instances of driving Under the Influence of Alcohol (DUI) was assessed with the following dichotomous item:

*Have you EVER had your driver's license or learner's permit suspended or revoked for moving violations related to alcohol use? (Yes/No)*

If respondents selected "yes" to the above question, they were asked the following question*:

*Have you EVER done things that could have easily hurt you or someone else like speeding or driving or using heavy machinery while under the influence of alcohol? (Yes/No)*

The first question was only presented to respondents that responded affirmative to having consumed alcohol.


* Note: For ease of interpretation, only the suspension question has been analysed.

```{r dui.incidence.tables, echo=F}
label(screened$DUI_suspended) <- "License Suspended or Revoked from DUI"

table1(~DUI_suspended, data = screened %>% filter(!is.na(DUI_suspended)), caption = "Suspended Overall")
table1(~DUI_suspended, data = screened %>% filter(meth_ingest == "Yes", !is.na(DUI_suspended)), caption = "Suspended MA Users")
table1(~DUI_suspended, data = screened %>% filter(meth_ingest == "No", !is.na(DUI_suspended)), caption = "Suspended Non-drug users")
```

```{r dui.incidence.plots, echo = F}
bar.dui.ma <- screened %>%
  filter(!is.na(DUI_suspended),
         meth_ingest == "Yes") %>% 
  ggplot(aes(x = DUI_suspended, fill = meth_ingest)) +
  geom_bar(aes(y = (..count..)/sum(..count..)*100), position = "dodge") +
  theme_light() +
  theme(legend.position = "none") +
  scale_fill_manual(values = .ma) +
  labs(x = "Licence Suspended", y = "Proportion (%)") +
  ggtitle("License Suspended from DUI (MA Users)") +
  coord_flip()

bar.dui.nma <- screened %>%
  filter(!is.na(DUI_suspended),
         meth_ingest == "No") %>% 
  ggplot(aes(x = DUI_suspended, fill = meth_ingest)) +
  geom_bar(aes(y = (..count..)/sum(..count..)*100), position = "dodge") +
  theme_light() +
  theme(legend.position = "none") +
  scale_fill_manual(values = .nma) +
  labs(x = "Licence Suspended", y = "Proportion (%)") +
  ggtitle("License Suspended from DUI (Non-drug Users)") +
  coord_flip()

grid.arrange(bar.dui.ma, bar.dui.nma, nrow = 2)

```

### Drugs (DUID)
Instances of driving under the influence of drugs (DUID) was assessed with the following question:

*Have you ever driven a car, motorcycle, truck, boat or other vehicle when you were under the influence of drugs?*

An affirmative response would then present the participant with questions relating to the recency of the DUID (Last 12 months and/or Before last 12 months) and the medicines or drugs that the DUID occurred with.  

Respondents were also presented with the same questions (regarding license suspension) as the DUI questions (TBC)

```{r DUID.prep, include = F}
screened <- screened %>% mutate(
  DUID_recent_meth = ifelse(str_detect(screened$DUID_recent_drugs, "Amphetamine"), TRUE, FALSE),
  DUID_lifetime_meth = ifelse(str_detect(screened$DUID_lifetime_drugs, "Amphetamine"), TRUE, FALSE),
  DUID_any_meth = ifelse(DUID_recent_meth == TRUE | DUID_lifetime_meth == TRUE, TRUE, FALSE),
  DUID_suspended_meth = ifelse(str_detect(screened$DUID_suspended_drugs, "Amphetamine"), TRUE, FALSE)
)
```


```{r DUID.tables, echo = F}
label(screened$DUID_incidence) <- "DUID Incidence"
label(screened$DUID_any_meth) <- "Ever Driven under the Influence of MA?"

table1(~DUID_incidence + DUID_any_meth + DUID_recent + DUID_recent_meth + DUID_lifetime + DUID_lifetime_meth + DUID_suspended + DUID_suspended_meth + DUID_ongoing, data = screened %>% filter(meth_ingest == "Yes"))
```

***

## Distribution of key variables by MA Use
### Dula Dangerous Driving Index (DDDI)
#### About {-}
The DDDI includes 28$^*$ items (divided into 3 categorical subscales) with each item assessing a persons tendency to engage in certain driving practices or beliefs held around specific driving behaviours.  Each item is scored on a 5-point Likert scale from 1 (Never) to 5 (Always).

<br>

The total score can be used as a composite measure for one's tendency and/or willingness to engage in dangerous driving.  Scores can range from 27 to 135.

<br>

#### Subscales: {-}

**Aggressive Driving:**
Assessed with 7 items which reflect behaviours that are intentionally meant to annoy, irritate or punish other drivers.  E.g. *I flash my headlights when I am annoyed by another driver.*

Scores range from 7 to 35.

<br>

**Negative Emotional Driving:**
Assessed with 9 items which gauge irritability and anger while driving and the tendency to become annoyed with other drivers.  E.g. *I lose my temper while driving.*

Scores range from 9 to 45.

<br>

**Risky Driving:**
Assessed with 11 items which gauge ones willingness to engage in unsafe driving behaviours.  E.g. *I will "drag race" other drivers at stop lights to get out front.*

Scores range from 11 to 55. 

<br>
<br>

$^*$The present study used 27 items with one item from the "Risky Driving" subscale - *I will weave in and out of slower traffic* - omitted from the questionnaire.  To view all the items used in the questionnaire, see the Appendix

***

#### Total Score {-}
```{r di.total, echo = F}
# Summary Table

# Histogram
hist.di <- screened %>% 
  ggplot(aes(x = DDDI_Total)) +
  geom_histogram(binwidth = 3, col = "black", fill = .grey) +
  labs(x = "DDDI Total Score", y = "") +
  theme_light()

# Hist by MA
hist.di.ma <- screened %>% 
  ggplot(aes(x = DDDI_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 3, col = "black") +
  labs(x = "DDDI Total Score", y = "", fill = "MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.di.ma <- screened %>% 
  ggplot(aes(x = DDDI_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "DDDI Total Score", y = "Density", fill = "MA Use") +
  theme_light()

# Box and scatter
box.di.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "DDDI Total") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.di, hist.di.ma, dens.di.ma, box.di.ma, nrow = 2, ncol = 2, top = "Distribution of Total DDDI Scores")

```

***

#### Aggressive Driving {-}
```{r di.AG, echo = F}
# Histogram
hist.ag <- screened %>% 
  ggplot(aes(x = DDDI_AD_Total)) +
  geom_histogram(binwidth = 2, col = "black", fill = .grey) +
  labs(x = "DDDI Aggression Score", y = "") +
  #ggtitle("Distribution of DDDI Total Score") +
  theme_light()

# Hist by MA
hist.ag.ma <- screened %>% 
  ggplot(aes(x = DDDI_AD_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 2, col = "black") +
  labs(x = "DDDI Aggression Score", y = "", fill = "MA Use") +
  #ggtitle("DDDI Aggression by MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.ag.ma <- screened %>% 
  ggplot(aes(x = DDDI_AD_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "DDDI Aggression Score", y = "Density", fill = "MA Use") +
  #ggtitle("DDDI Aggression by MA Use (Density)") +
  theme_light()

# Box and scatter
box.ag.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "DDDI Aggression") +
  #ggtitle("Box plot of DDDI Aggression by MA Use") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.ag, hist.ag.ma, dens.ag.ma, box.ag.ma, nrow = 2, top = "Distribution of DDDI Aggressive Driving Scores")
```

***

#### Negative Emotional Driving {-}
```{r di.NE, echo = F}
# Histogram
hist.ned <- screened %>% 
  ggplot(aes(x = DDDI_NE_Total)) +
  geom_histogram(binwidth = 2, col = "black", fill = .grey) +
  labs(x = "DDDI NED Score", y = "") +
  #ggtitle("Distribution of DDDI Negative Emotional Driving (NED) Score") +
  theme_light()

# Hist by MA
hist.ned.ma <- screened %>% 
  ggplot(aes(x = DDDI_NE_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 2, col = "black") +
  labs(x = "DDDI NED Score", y = "", fill = "MA Use") +
  #ggtitle("DDDI NED by MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.ned.ma <- screened %>% 
  ggplot(aes(x = DDDI_NE_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "DDDI NED Score", y = "Density", fill = "MA Use") +
  #ggtitle("DDDI NED by MA Use (Density)") +
  theme_light()

# Box and scatter
box.ned.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "DDDI NED") +
  #ggtitle("Box plot of DDDI NED by MA Use") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.ned, hist.ned.ma, dens.ned.ma, box.ned.ma, nrow = 2, top = "Distribution of DDDI Negative Emotional Driving Scores")

```

***

#### Risky Driving {-}
```{r di.RD, echo = F}
# Histogram
hist.rd <- screened %>% 
  ggplot(aes(x = DDDI_RD_Total)) +
  geom_histogram(binwidth = 2, col = "black", fill = .grey) +
  labs(x = "DDDI Risky Driving Score", y = "") +
  #ggtitle("Distribution of DDDI Risky Driving Score") +
  theme_light()

# Hist by MA
hist.rd.ma <- screened %>% 
  ggplot(aes(x = DDDI_RD_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 2, col = "black") +
  labs(x = "DDDI Risky Driving Score", y = "", fill = "MA Use") +
  #ggtitle("DDDI Risky Driving by MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.rd.ma <- screened %>% 
  ggplot(aes(x = DDDI_RD_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "DDDI Risky Driving Score", y = "Density", fill = "MA Use") +
  #ggtitle("DDDI Risky Driving by MA Use (Density)") +
  theme_light()

# Box and scatter
box.rd.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "DDDI Risky Driving") +
  #ggtitle("Box plot of DDDI Risky Driving by MA Use") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.rd, hist.rd.ma, dens.rd.ma, box.rd.ma, nrow = 2, top = "Distribution of DDDI Risky Driving Scores")

```

***

#### Aggressive Driving Items {-}
```{r di.ad.ind, echo = F}
box.ad.1 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_1, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("I flash my headlights when I am annoyed by another driver")

box.ad.2 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_2, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("I make rude gestures toward drivers who annoy me")

box.ad.3 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_3, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("I verbally insult drivers who annoy me")

box.ad.4 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_4, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "Score") +
  theme(legend.position = "none") +
  ggtitle("I would tailgate a driver who annoys me")

box.ad.5 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_5, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("I deliberately use my car/truck to block drivers who tailgate me.")

box.ad.6 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_6, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("I feel it is my right to strike back in some way...")

box.ad.7 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_AD_7, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.2, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("When someone cuts me off, I feel I should punish him/her")

grid.arrange(box.ad.1, box.ad.2, box.ad.3, box.ad.4) 
grid.arrange(box.ad.5, box.ad.6, box.ad.7, ncol = 2)

```


#### Negative Emotional Driving Items {-}
```{r di.ned.ind, echo = F}
box.ne.1 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_1, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q65")

box.ne.2 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_2, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q66")

box.ne.3 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_3, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q67")

box.ne.4 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_4, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q79")

box.ne.5 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_5, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q83")

box.ne.6 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_6, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q84")

box.ne.7 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_7, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q85")

box.ne.8 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_8, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q88")

box.ne.9 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_NE_9, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q92")

grid.arrange(box.ne.1, box.ne.2, box.ne.3, box.ne.4, box.ne.5, ncol = 2)
grid.arrange(box.ne.6, box.ne.7, box.ne.8, box.ne.9, ncol = 2)

```

#### Risky Driving Items {-}
```{r di.rd.ind, echo = F}
box.rd.1 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_1, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q74")

box.rd.2 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_2, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q76")

box.rd.3 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_3, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q80")

box.rd.4 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_4, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q81")

box.rd.5 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_5, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q86")

box.rd.6 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_6, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q87")

box.rd.7 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_7, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q89")

box.rd.8 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_8, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q90")

box.rd.9 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_9, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q92")

box.rd.10 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_10, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q93")

box.rd.11 <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DDDI_RD_11, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, height = 0, alpha = 0.5) +
  theme_light() +
  theme(legend.position = "none") +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Q94")

grid.arrange(box.rd.1, box.rd.2, box.rd.3, box.rd.4, box.rd.5, box.rd.6)
grid.arrange(box.rd.7, box.rd.8, box.rd.9, box.rd.10, box.rd.11)
```

***

### STAXI-2
#### About {-}
The State-Trait Anger Expression Inventory (STAXI-2) is designed to understand the attitudes, emotional states and behaviours that are typical to a person in regard to their feelings, control and expression of anger.  It distinguishes anger between being a passing/present state as a personality trait.

The full STAXI-2 contains 57 items grouped into 12 subscales which lie under three factors (State Anger, Trait Anger and Anger Control and Expression).  

For the purposes of this study, only the subscacles relating to State and Trait Anger were used.  

##### Scoring {-}
Each Item on the STAXI-2 is rated on a 4-point Likert scale from 1 (Not al all/Almost Never) to 4 (Very much so/Almost Always).
Total score for State or Trait Anger is obtained by adding up the items from each subscale.  Total scores for the subscales are obtained by adding up the scores of each item for the corresponding subscale.  

#### State Anger {-}
Assesses the intensity of the individual's angry feelings at time of testing.

Minimum Possible Score = 14.  Maximum Possible Score = 56.

##### Subscales: {-}
**Feeling Angry (S-Ang/F):**
Measures the intensity of the angry feelings the person may be experiencing.

<br>

**Feel like verbally expressing anger (S-Ang/V):**
Measures the intensity of the current feelings related to the verbal expression of anger.

<br>

**Feel like physically expressing anger (A-Ang/P):**
Measures the intensity of the current feelings related to the physical expression of anger.

<br>

#### Trait Anger {-}
Measures the general tendency of a person to get angry. 

Minimum Possible Score = 10.  Maximum Possible Score = 40.

##### Subscales: {-}
**Angry Temperament (T-Ang/T):**
Measures the disposition to experience anger without specific provocation. 

<br>

**Angry Reaction (T-Ang/T):**
Measures the frequency with which angry feelings are experienced in situations that involve frustration and/or negative evaluations.  
<br>


***

#### State Anger Total {-}
```{r staxi.state, echo = F}
# Histogram
hist.sstaxi <- screened %>% 
  ggplot(aes(x = S_Ang_Total)) +
  geom_histogram(binwidth = 1, col = "black", fill = .grey) +
  labs(x = "State Anger", y = "") +
  #ggtitle("Distribution of State Anger Scores") +
  theme_light()

# Hist by MA
hist.sstaxi.ma <- screened %>% 
  ggplot(aes(x = S_Ang_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 1, col = "black") +
  labs(x = "State Anger", y = "", fill = "MA Use") +
  #ggtitle("State Anger by MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.sstaxi.ma <- screened %>% 
  ggplot(aes(x = S_Ang_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "State Anger", y = "Density", fill = "MA Use") +
  #ggtitle("State Anger by MA Use (Density)") +
  theme_light()

# Box and scatter
box.sstaxi.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = S_Ang_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "State Anger") +
  #ggtitle("Box plot of State Anger by MA Use") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.sstaxi, hist.sstaxi.ma, dens.sstaxi.ma, box.sstaxi.ma, ncol = 2, top = "Distribution of State STAXI Scores (Total)")
```

#### State anger Subscales {-}
```{r staxi.state.subs, echo=F}
# Feeling box
box.sstaxi.F <- screened %>% 
  ggplot(aes(x = meth_ingest, y = S_Ang_F_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Feeling") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

#Verbal box
box.sstaxi.V <- screened %>% 
  ggplot(aes(x = meth_ingest, y = S_Ang_V_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "") +
  ggtitle("Verbal") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Physical box
box.sstaxi.P <- screened %>% 
  ggplot(aes(x = meth_ingest, y = S_Ang_P_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "") +
  ggtitle("Physical") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Feeling Dens
dens.sstaxi.F <- screened %>% 
  ggplot(aes(x = S_Ang_F_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "Feeling", y = "Density", fill = "MA Use") +
  theme_light() +
  theme(legend.position = "none")

# Verbal Dens
dens.sstaxi.V <- screened %>% 
  ggplot(aes(x = S_Ang_V_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "Verbal", y = "", fill = "MA Use") +
  theme_light() +
  theme(legend.position = "none")

# Physical Dens
dens.sstaxi.P <- screened %>% 
  ggplot(aes(x = S_Ang_P_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "Physical", y = "", fill = "MA Use") +
  theme_light() +
  theme(legend.position = "none")

grid.arrange(box.sstaxi.F, box.sstaxi.V, box.sstaxi.P, 
             dens.sstaxi.F, dens.sstaxi.V, dens.sstaxi.P, nrow = 2, ncol = 3)
```

***

#### Trait Anger Total {-}
```{r staxi.trait, echo = F}
# Histogram
hist.tstaxi <- screened %>% 
  ggplot(aes(x = T_Ang_Total)) +
  geom_histogram(binwidth = 1, col = "black", fill = .grey) +
  labs(x = "Trait Anger", y = "") +
  #ggtitle("Distribution of Trait Anger Scores") +
  theme_light()

# Hist by MA
hist.tstaxi.ma <- screened %>% 
  ggplot(aes(x = T_Ang_Total, fill = meth_ingest)) +
  geom_histogram(binwidth = 1, col = "black") +
  labs(x = "Trait Anger", y = "", fill = "MA Use") +
  #ggtitle("Trait Anger by MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.tstaxi.ma <- screened %>% 
  ggplot(aes(x = T_Ang_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "Trait Anger", y = "Density", fill = "MA Use") +
  #ggtitle("Trait Anger by MA Use (Density)") +
  theme_light()

# Box and scatter
box.tstaxi.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = T_Ang_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.1, height = 0, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "Trait Anger") +
  #ggtitle("Box plot of Trait Anger by MA Use") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.tstaxi, hist.tstaxi.ma, dens.tstaxi.ma, box.tstaxi.ma, ncol = 2, top = "Distribution of Trait Staxi Scores (Total)")

```

#### Trait Anger Subscales {-}
```{r staxi.trait.subs, echo = F}
# Temperament box
box.tstaxi.T <- screened %>% 
  ggplot(aes(x = meth_ingest, y = T_Ang_T_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "Score") +
  ggtitle("Angry Temperament") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Reaction box
box.tstaxi.R <- screened %>% 
  ggplot(aes(x = meth_ingest, y = T_Ang_R_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(width = 0.08, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "") +
  ggtitle("Angry Reaction") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Temperament dens
dens.tstaxi.T <- screened %>% 
  ggplot(aes(x = T_Ang_T_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "Temperament", y = "Density", fill = "MA Use") +
  theme_light() +
  theme(legend.position = "none")

# Reaction dens
dens.tstaxi.R <- screened %>% 
  ggplot(aes(x = T_Ang_R_Total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "Reaction", y = "", fill = "MA Use") +
  theme_light() +
  theme(legend.position = "none")

grid.arrange(box.tstaxi.T, box.tstaxi.R, 
             dens.tstaxi.T, dens.tstaxi.R, nrow = 2, ncol = 2)

```

***

### K6 Distress Scale
The K6 Distress scale is a 6-item assessment that measures overall distress from depressive and anxiety related symptoms.

It measures distress for the period 4 weeks prior to administration of the test.  

Each Item is rated on a 5-point$^*$ likert scale ranging from 1 (none of the time) to 5 (all of the time).

Total score is obtained by simply adding up the item scores together.

The minimum score one can obtain is 6 and the maximum is 30.


$^*$The present study used a 4-point scale for the item *During the past 30 days, about how often did you feel depressed?* with the level "A little of the time" omitted from the survey.

#### Numeric summary {-}
```{r num.summ.k6, echo=F}
label(screened$k6_total) <- "K6 Distress Score"

table1(~k6_total, data = screened, caption = "Overall K6 Distress Scores")

table1(~k6_total, data = screened %>% filter(meth_ingest == "Yes"), caption = "K6 Distress Scores for MA Users")

table1(~k6_total, data = screened %>% filter(meth_ingest != "Yes"), caption = "K6 Distress Scores for non-MA Users")
```


#### K6 Distribution {-}
```{r dist.k6, echo = F, warning=F}
# Hist
hist.k6 <- screened %>% ggplot(aes(x = k6_total)) +
  geom_histogram(binwidth = 1, col = "black", fill = .grey) +
  labs(x = "K6 Distress Score", y = "") +
  theme_light();

# Hist by MA
hist.k6.ma <- screened %>% 
  ggplot(aes(x = k6_total, fill = meth_ingest)) +
  geom_histogram(binwidth = 1, col = "black") +
  labs(x = "K6 Distress Score", y = "", fill = "MA Use") +
  theme_light() +
  facet_grid(meth_ingest ~.) +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

# Dens by MA
dens.k6.ma <- screened %>% 
  ggplot(aes(x = k6_total, fill = meth_ingest)) +
  geom_density(col = "black", alpha = 0.5) +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "K6 Distress Score", y = "Density", fill = "MA Use") +
  theme_light()

# Boxplot by MA
box.k6.ma <- screened %>% 
  ggplot(aes(x = meth_ingest, y = k6_total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, alpha = 0.8) +
  geom_jitter(width = 0.1, height = 0, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "K6 Distress Score") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(hist.k6, hist.k6.ma, dens.k6.ma, box.k6.ma, top = "Distribution of K6 Distress Scores")
```

### Severity Dependence Scale (SDS)
The SDS is a 5 item questionnaire that provides a score indicating the severity of dependence on a given substance.
Each item is scored from 0 to 3, with a higher score indicating greater dependence.

Those whose total score is above 4 are considered to be dependent users.

```{r sds, echo=F}
ma <- screened %>% filter(meth_ingest == "Yes")
label(ma$meth_SDS) <- "SDS"
table1(~meth_SDS, data = ma)

ma %>% 
  ggplot(aes(x = meth_SDS)) +
  geom_histogram(binwidth = 1, col = "black", fill = .grey) +
  theme_light() +
  labs(x = "Severity Dependence Score", y = "") +
  ggtitle("Distribution of SDS Scores")
```

***

### Attitudes
Two questionnaires designed to assess participants attitudes toward drink and drug driving were administered.  
Each questionnaire included 10 items rated on a 7-point Likert scale indicating to what extend the participant agreed or disagreed with the statement (e.g. *My community needs stricter laws against drink driving*).  After taking into account items that were reverse scored (e.g. *It's OK to drive after drinking as long as you're not too drunk*), total scores were calculated with higher scores indicating a more favourable attitude toward drink or drug driving.

#### Attitudes toward Drink Driving (DUI) {-}
```{r attitudes.DUI, echo = F}
label(screened$DUI_attitude_Total) <- "DUI Attitude Total"
table1(~DUI_attitude_Total, data = screened, caption = "DUI Attitude")
table1(~DUI_attitude_Total, data = screened %>% filter(meth_ingest == "Yes"), caption = "DUI Attitude MA Users")
table1(~DUI_attitude_Total, data = screened %>% filter(meth_ingest == "No"), caption = "DUI Attitude non-drug Users")

dens.dui.att <- screened %>% 
  ggplot(aes(x = DUI_attitude_Total, fill = meth_ingest)) +
  geom_density(alpha = 0.5, col = "black") +
  theme_light() +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "DUI Attitude Score", y = "Density", fill = "MA User")

box.dui.att <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DUI_attitude_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(height = 0, width = 0.1, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "DUI Attitude Score", fill = "MA User") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(dens.dui.att, box.dui.att, nrow = 2)

```


#### Attitudes toward Drug Driving {-}



```{r attitudes.DUID, echo = F, warning=F}
label(screened$DUID_attitude_Total) <- "DUID Attitude Total"
table1(~DUID_attitude_Total, data = screened, caption = "DUID Attitude")
table1(~DUID_attitude_Total, data = screened %>% filter(meth_ingest == "Yes"), caption = "DUID Attitude MA Users")
table1(~DUID_attitude_Total, data = screened %>% filter(meth_ingest == "No"), caption = "DUID Attitude non-drug Users")

dens.duid.att <- screened %>% 
  ggplot(aes(x = DUID_attitude_Total, fill = meth_ingest)) +
  geom_density(alpha = 0.5, col = "black") +
  theme_light() +
  scale_fill_manual(values = c(.ma, .nma)) +
  labs(x = "DUID Attitude Score", y = "Density", fill = "MA User")

box.duid.att <- screened %>% 
  ggplot(aes(x = meth_ingest, y = DUID_attitude_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(height = 0, width = 0.1, alpha = 0.5) +
  theme_light() +
  labs(x = "MA Use", y = "DUID Attitude Score", fill = "MA User") +
  scale_fill_manual(values = c(.ma, .nma))

grid.arrange(dens.duid.att, box.duid.att, nrow = 2)
```
