==> 03-bivariate.Rmd <==
# Bivariate Descriptives {#Bivar.Descriptives}
```{r library, include = FALSE}
library(bookdown)
library(tidyverse)
library(gridExtra)
library(rlang) # For ggplot functions
library(knitr)
library(kableExtra)
```

```{r data.load, include = F}
load("data/objects/data.full.Rda")
load("data/objects/screened.full.Rda")
load("data/objects/screened.Rda")
```

```{r data.preparation, include = F}
# Preparing colours for graphs
.grey <- "#A6A6A6"
.ma <- "#DC2D27"
.nma <- "#39B9FF"
.lav <- "#D9BBDC"
.lgreen <- "#BEDCB3"

# Changing NA from meth_ingest and meth_user to "No"
screened$meth_ingest[is.na(screened$meth_ingest)] <- "No"
screened$meth_user[is.na(screened$meth_user)] <- "Non-User"
screened$meth_user <- as.factor(screened$meth_user)

# Rearranging levels for education 
screened$education <- factor(screened$education, levels = c("Did not finish high school",
                                                          "High School Diploma",
                                                          "Vocational/Technical degree or certificate",
                                                          "Did not finish University",
                                                          "Bachelor Degree",
                                                          "Postgraduate Degree"))
```

This section will explore bivariate relationships between each key predictor variable (STAXI-2, K6 Distress, SDS, Attitudes) and will examine in more detail their relationship with the outcome variable (DDDI).  There will be particular attention paid on the relationship between the predictors and the outcome variable and we will seek to graphically identify whether there may be an y interactions present between the predictor variables and MA use on the DDDI.
```{r bivar.plot.function, include=FALSE}
plot.bivar <- function(df, xcol, ycol, colour) {
  ggplot(data = df, aes(x = {{xcol}}, y = {{ycol}}, col = {{colour}})) +
    geom_smooth(method = "lm", alpha = 0.1, fullrange = T) +
    geom_point(size = 2, alpha = 0.5) +
    theme_light() +
    scale_color_manual(values = c(.ma, .nma))
}
```
## STAXI-2

### State Anger (STAXI-S)
#### DDDI by STAXI-S (ignoring MA) {-}
```{r staxi.state.bi, echo = F, message=F, warning=F}
plot.bivar(screened, S_Ang_Total, DDDI_Total) +
  labs(x = "State Anger Total", y = "DDDI Total") +
  ggtitle("DDDI by State Anger")
```


#### DDDI by STAXI-S (Stratified by MA) {-}

```{r staxi.state.bivar, echo = F, message=F, warning=F}
point.sstaxi <- plot.bivar(screened, S_Ang_Total, DDDI_Total, meth_ingest) +
  labs(x = "State Total", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by State Anger")

point.sstaxi.F <- plot.bivar(screened, S_Ang_F_Total, DDDI_Total, meth_ingest) +
  labs(x = "Feeling", y = "", col = "MA Use") +
  theme(legend.position = "none")

point.sstaxi.V <- plot.bivar(screened, S_Ang_V_Total, DDDI_Total, meth_ingest) +
  labs(x = "Verbal", y = "", col = "MA Use") +
  theme(legend.position = "none")

point.sstaxi.P <- plot.bivar(screened, S_Ang_P_Total, DDDI_Total, meth_ingest) +
  labs(x = "Physical", y = "", col = "MA Use") +
  theme(legend.position = "none")

lay <- rbind(c(2, 3, 4),
             c(2, 3, 4))

point.sstaxi

grid.arrange(point.sstaxi.F, point.sstaxi.V, point.sstaxi.P, layout_matrix = lay, top = "STAXI-S Subscales")
```

### Trait Anger (STAXI-T)
#### DDDI by STAXI-T (ignoring MA) {-}
```{r staxi.trait.bi, echo = F, message = F, warning = F}
plot.bivar(screened, T_Ang_Total, DDDI_Total) +
  labs(x = "Trait Anger Total", y = "DDDI Total") +
  ggtitle("DDDI by Trait Anger (ignoring MA use)")

```

#### DDDI by STAXI-T (stratified by MA Use) {-}
```{r staxi.trait.bivar, echo = F, message = F, warning = F}
point.tstaxi <- plot.bivar(screened, T_Ang_Total, DDDI_Total, meth_ingest) +
  labs(x = "Trait ", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by Trait Anger (stratified by MA use)")

point.tstaxi.R <- plot.bivar(screened, T_Ang_R_Total, DDDI_Total, meth_ingest) +
  labs(x = "Reaction",y = "", col = "MA Use") +
  theme(legend.position = "none")

point.tstaxi.T <- plot.bivar(screened, T_Ang_T_Total, DDDI_Total, meth_ingest) +
  labs(x = "Temperament", y = "", col = "MA Use") +
  theme(legend.position = "none")

lay <- rbind(c(2, 2, 3, 3),
             c(2, 2, 3, 3))

point.tstaxi
grid.arrange(point.tstaxi.R, point.tstaxi.T, layout_matrix = lay, top = "STAXI-T Subscales")
```

#### DDDI Subscales by STAXI-T (stratified by MA Use) {-}
```{r staxi.trait.dddi.subscales, echo = F, message = F, warning = F}
point.tstaxi.AD <- plot.bivar(screened, T_Ang_Total, DDDI_AD_Total, meth_ingest) +
  labs(x = "Trait Anger Total", y = "") +
  ggtitle("Aggressive Driving") +
  theme(legend.position = "none")

point.tstaxi.NE <- plot.bivar(screened, T_Ang_Total, DDDI_NE_Total, meth_ingest) +
  labs(x = "Trait Anger Total", y = "") +
  ggtitle("Negative Emotional Driving") +
  theme(legend.position = "none")

point.tstaxi.RD <- plot.bivar(screened, T_Ang_Total, DDDI_RD_Total, meth_ingest) +
  labs(x = "Trait Anger Total", y = "") +
  ggtitle("Risky Driving") +
  theme(legend.position = "none")

grid.arrange(point.tstaxi.AD, point.tstaxi.NE, point.tstaxi.RD, ncol = 3)

```

***

## K6 Distress Scale
```{r k6.bivar, echo=F, message = F, warning = F}
plot.bivar(screened, k6_total, DDDI_Total) +
  labs(x = "K6 Distress Score", y = "DDDI Total")

plot.bivar(screened, k6_total, DDDI_Total, meth_ingest) +
  labs(x = "K6 Distress Score", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by K6 Distress Score (Stratified by MA use")
```

***

## SDS Scale
Bivariate relationship between SDS and DDDI (MA Users only)
```{r SDS.bivar, echo=F, message=F, warning=F}
plot.bivar(screened, meth_SDS, DDDI_Total)

point.di.sds <- plot.bivar(screened, meth_SDS, DDDI_Total) +
  labs(x = "SDS Score", y = "DDDI Score") +
  theme(legend.position = "none") +
  ggtitle("DDDI total")

point.di.ad.sds <- plot.bivar(screened, meth_SDS, DDDI_AD_Total) +
  labs(x = "SDS Score", y = "") +
  ggtitle("Aggressive Driving")

point.di.ne.sds <- plot.bivar(screened, meth_SDS, DDDI_NE_Total) +
  labs(x = "SDS Score", y = "DDDI Negative Emotional Driving") +
  theme(legend.position = "none") +
  ggtitle("Negative Emotional Driving")

point.di.rd.sds <- plot.bivar(screened, meth_SDS, DDDI_RD_Total) +
  labs(x = "SDS Score", y = "") +
  ggtitle("Risky Driving")

grid.arrange(point.di.sds, point.di.ad.sds, point.di.ne.sds, point.di.rd.sds, ncol = 2)
```

***

## Drug Use
### Alcohol (AUDIT-C)
```{r alcohol.bivar, echo = F, warning=F, message=F}
point.di.audit <- plot.bivar(screened, AUDIT_Total, DDDI_Total) +
  labs(x = "AUDIT-C Total", y = "DDDI Total") +
  ggtitle("DDDI Total by AUDIT-C")

point.di.audit.ma <- plot.bivar(screened, AUDIT_Total, DDDI_Total, meth_ingest) +
  labs(x = "AUDIT-C Total", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI Total by AUDIT-C (Stratified by MA)")


box.audit.di <- screened %>% 
  ggplot(aes(x = AUDIT_Risky, y = DDDI_Total)) +
  geom_boxplot(outlier.alpha = 0, height = 0, width = 0.3, alpha = 0.5) +
  geom_jitter(height = 0, width = 0.05, alpha = 0.8) +
  theme_light() +
  labs(x = "AUDIT-C Risky Drinking", y = "DDDI Total") +
  ggtitle("DDDI Total by AUDIT-C Risky Drinking")

box.audit.di.ma <- screened %>% 
  ggplot(aes(x = AUDIT_Risky, y = DDDI_Total, fill = meth_ingest)) +
  geom_boxplot(outlier.alpha = 0, height = 0, width = 0.3, alpha = 0.5) +
  geom_jitter(aes(col = meth_ingest), height = 0, width = 0.05, alpha = 0.8) +
  theme_light() +
  labs(x = "AUDIT-C Risky Drinking", y = "DDDI Total", fill = "MA Use", col = "MA Use") +
  ggtitle("DDDI Total by AUDIT-C Risky Drinking (Stratified by MA)")

grid.arrange(point.di.audit, point.di.audit.ma)
grid.arrange(box.audit.di, box.audit.di.ma, nrow = 2)
```

```{r drug.use.bivar, echo = F}


```

***

## Age
```{r age bivar, echo=F}
plot.bivar(screened, age, DDDI_Total) +
  labs(x = "Age", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by Age")

plot.bivar(screened, age, DDDI_Total, meth_ingest) +
  labs(x = "Age", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by Age (stratified by MA use)")

# screened %>% 
#   ggplot(aes(x = age, y = DDDI_Total, col = sex)) +
#   geom_point(size = 2) +
#   theme_light() +
#   labs(x = "Age", y = "DDDI Total Score") +
#   facet_grid(~meth_ingest) +
#   ggtitle("Does this work")
```

***

## Attitudes
### DUI
```{r DUI, echo = F, message = F, warning = F}
plot.bivar(screened, DUI_attitude_Total, DDDI_Total, meth_ingest) +
  labs(x = "DUI Attitude Total", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by DUI Attitude")
```

```{r DUID, echo = F, message = F, warning = F}
plot.bivar(screened, DUID_attitude_Total, DDDI_Total, meth_ingest) +
  labs(x = "DUID Attitude Total", y = "DDDI Total", col = "MA Use") +
  ggtitle("DDDI by DUID Attitude")
```

## Correlation Matrix
```{r cor.matrix.functionn, include=F}
#' correlation_matrix from https://www.r-bloggers.com/2020/07/create-a-publication-ready-correlation-matrix-with-significance-levels-in-r/
#' Creates a publication-ready / formatted correlation matrix, using `Hmisc::rcorr` in the backend.
#'
#' @param df dataframe; containing numeric and/or logical columns to calculate correlations for
#' @param type character; specifies the type of correlations to compute; gets passed to `Hmisc::rcorr`; options are `"pearson"` or `"spearman"`; defaults to `"pearson"`
#' @param digits integer/double; number of decimals to show in the correlation matrix; gets passed to `formatC`; defaults to `3`
#' @param decimal.mark character; which decimal.mark to use; gets passed to `formatC`; defaults to `.`
#' @param use character; which part of the correlation matrix to display; options are `"all"`, `"upper"`, `"lower"`; defaults to `"all"`
#' @param show_significance boolean; whether to add `*` to represent the significance levels for the correlations; defaults to `TRUE`
#' @param replace_diagonal boolean; whether to replace the correlations on the diagonal; defaults to `FALSE`
#' @param replacement character; what to replace the diagonal and/or upper/lower triangles with; defaults to `""` (empty string)
#'
#' @return a correlation matrix
#' @export
#'
#' @examples
#' `correlation_matrix(iris)`
#' `correlation_matrix(mtcars)`
correlation_matrix <- function(df, 
                               type = "pearson",
                               digits = 3, 
                               decimal.mark = ".",
                               use = "all", 
                               show_significance = TRUE, 
                               replace_diagonal = FALSE, 
                               replacement = ""){
  
  # check arguments
  stopifnot({
    is.numeric(digits)
    digits >= 0
    use %in% c("all", "upper", "lower")
    is.logical(replace_diagonal)
    is.logical(show_significance)
    is.character(replacement)
  })
  # we need the Hmisc package for this
  require(Hmisc)
  
  # retain only numeric and boolean columns
  isNumericOrBoolean = vapply(df, function(x) is.numeric(x) | is.logical(x), logical(1))
  if (sum(!isNumericOrBoolean) > 0) {
    cat('Dropping non-numeric/-boolean column(s):', paste(names(isNumericOrBoolean)[!isNumericOrBoolean], collapse = ', '), '\n\n')
  }
  df = df[isNumericOrBoolean]
  
  # transform input data frame to matrix
  x <- as.matrix(df)
  
  # run correlation analysis using Hmisc package
  correlation_matrix <- Hmisc::rcorr(x, type = )
  R <- correlation_matrix$r # Matrix of correlation coeficients
  p <- correlation_matrix$P # Matrix of p-value 
  
  # transform correlations to specific character format
  Rformatted = formatC(R, format = 'f', digits = digits, decimal.mark = decimal.mark)
  
  # if there are any negative numbers, we want to put a space before the positives to align all
  if (sum(R < 0) > 0) {
    Rformatted = ifelse(R > 0, paste0(' ', Rformatted), Rformatted)
  }
  
  # add significance levels if desired
  if (show_significance) {
    # define notions for significance levels; spacing is important.
    stars <- ifelse(is.na(p), "   ", ifelse(p < .001, "***", ifelse(p < .01, "** ", ifelse(p < .05, "*  ", "   "))))
    Rformatted = paste0(Rformatted, stars)
  }
  # build a new matrix that includes the formatted correlations and their significance stars
  Rnew <- matrix(Rformatted, ncol = ncol(x))
  rownames(Rnew) <- colnames(x)
  colnames(Rnew) <- paste(colnames(x), "", sep =" ")
  
  # replace undesired values
  if (use == 'upper') {
    Rnew[lower.tri(Rnew, diag = replace_diagonal)] <- replacement
  } else if (use == 'lower') {
    Rnew[upper.tri(Rnew, diag = replace_diagonal)] <- replacement
  } else if (replace_diagonal) {
    diag(Rnew) <- replacement
  }
  
  return(Rnew)
}
```

```{r cor.matrix.biv, echo=F, message = F}
key.vars <- screened %>%
  select(DDDI_Total, DDDI_AD_Total, DDDI_RD_Total, DDDI_NE_Total, 
         k6_total, 
         S_Ang_Total, S_Ang_F_Total, S_Ang_V_Total, S_Ang_P_Total,
         T_Ang_Total, T_Ang_R_Total, T_Ang_T_Total,
         DUI_attitude_Total,
         DUID_attitude_Total)

key.vars.total <- screened %>%
  select(DDDI_Total, DDDI_AD_Total, DDDI_RD_Total, DDDI_NE_Total,
         age,
         k6_total, 
         S_Ang_Total,
         T_Ang_Total, T_Ang_R_Total, T_Ang_T_Total,
         AUDIT_Total,
         DUI_attitude_Total,
         DUID_attitude_Total)

corr.vars <- correlation_matrix(key.vars.total, digits = 2, use = "lower")

kable(corr.vars) %>% 
  kable_styling("striped") %>% 
  scroll_box(width = "100%")
```

```{r cor.scatter, echo = F, message = F}

```
