==> 05-additional.Rmd <==
# Additional Analyses
```{r additional.lib, include=F}
library(bookdown)
library(tidyverse)
library(apaTables)
library(gridExtra)
library(table1)
library(kableExtra)
library(car) #VIF
# library(parameters) # Standardised estimates
library(olsrr)
```

```{r data.load.add, include = F}
load("data/objects/screened.full.Rda")
load("data/objects/screened.Rda")
```

```{r data.preparation.add, include = F}
# Preparing colours for graphs
.grey <- "#A6A6A6"
.ma <- "#DC2D27"
.nma <- "#39B9FF"
.lav <- "#D9BBDC"
.lgreen <- "#BEDCB3"

# Changing NA from meth_ingest and meth_user to "No"
screened$meth_ingest[is.na(screened$meth_ingest)] <- "No"
screened$meth_user[is.na(screened$meth_user)] <- "Non-User"
screened$meth_user <- as.factor(screened$meth_user)

# Rearranging levels for education 
screened$education <- factor(screened$education, levels = c("Did not finish high school",
                                                          "High School Diploma",
                                                          "Vocational/Technical degree or certificate",
                                                          "Did not finish University",
                                                          "Bachelor Degree",
                                                          "Postgraduate Degree"))

# Filtering for MA only
ma <- screened %>% filter(meth_ingest == "Yes")

# Converting non-drinkers to 0 on AUDIT-C
ma <- ma %>% 
  mutate(AUDIT_Total = ifelse(is.na(AUDIT_Total), 0, AUDIT_Total))

options(scipen = 999) # Removing scientific notation
```

## Reduced Model
### Model Building
```{r not-forced}
m0 <- lm(DDDI_Total~1, data = ma) # Total SS
m1 <- lm(DDDI_Total ~ age + sex, data = ma)
m2.rm <- lm(DDDI_Total ~ age + AUDIT_Total + meth_SDS, data = ma)
m3.rm <- lm(DDDI_Total ~ AUDIT_Total + meth_SDS + T_Ang_Total, data = ma)
save(m3.rm, file = "objects/m3.rm.rda")
```

### Sums of Suares
#### Total SS {-}
```{r total.ss.step, echo=F}
anova(m0) %>% kable(align = "c") %>% kable_styling()
```

#### Model SS {-}
```{r model.ss.step, echo=F}
anova(m1, m2.rm, m3.rm) %>% kable(align = "c") %>% kable_styling()
```
### Model Summaries
#### Model 1 {-}
```{r, echo = F}
summary(m1)
```

#### Model 2 {-}
```{r, echo=F}
summary(m2.rm)
```

#### Model 3 {-}
```{r, echo=F}
ols_regress(m3.rm)
```
### Assumptions
#### Plot of Residuals {-}
```{r, echo=F}
std.res <- rstandard(m3.rm)
m3.rm.std.res <- tibble(
  Std.Res = std.res,
  Fitted = m3.rm$fitted.values
)

m3.rm.std.res %>% 
  ggplot(aes(x = Fitted, y = Std.Res)) +
  geom_point() +
  geom_hline(aes(yintercept = 0), linetype = "dashed", alpha = 0.5) +
  theme_light() +
  labs(x = "Fitted", y = "Standardised Residuals") +
  ggtitle("Plot of Residuals")
```

#### Colinearity {-}
```{r, echo=F}
tibble(Variable = c("AUDIT-C", "SDS", "STAXI-T"),
       VIF = vif(m3.rm))
```

#### Homogeneity {-}
```{r ass.homo.rm, echo=F}
plot(m3.rm,1)
```

#### Linearity {-}
```{r ass.lin.rm, echo=F}
plot(m3.rm, 5)
```

#### Normality of Residuals {-}
```{r ass.norm.rm, echo=F}
plot(m3.rm, 2)
```

#### Outliers {-}
##### Residuals {-}
```{r ass.res.std.dev, echo=F}
ma %>% 
  ggplot(aes(x = as.numeric(row.names(ma)), y = std.res)) +
  geom_point() +
  geom_hline(aes(yintercept = 2), linetype = "dashed", alpha = 0.5) +
  geom_hline(aes(yintercept = -2), linetype = "dashed", alpha = 0.5) +
  geom_text(size = 3,
            label = as.numeric(row.names(ma)),
            nudge_x = 0.25,
            nudge_y = 0.25,
            check_overlap = T) +
  theme_light() +
  labs(x = "Index", y = "Std Residuals") +
  ggtitle("Std. Residual by Index")
```

##### Cooks D {-}
```{r cooks.doodle, echo=F}
# Calculate Cook's D for each observation
cooksD <- cooks.distance(m3.rm)

# Plot Cook's D
n <- nrow(ma)

ma %>% 
  ggplot(aes(x = as.numeric(row.names(ma)), y = cooksD)) +
  geom_point() +
  geom_hline(yintercept = 4/n, linetype = "dashed") +
  geom_text(size = 2.5,
            label = as.numeric(row.names(ma)),
            nudge_y = 0.025,
            check_overlap = T) +
  theme_light() +
  labs(x = "Index", y = "Cooks Distance") +
  ggtitle("Cooks Distance")
```

##### Mahalanobis Distance {-}
```{r mahalinobis.rm, echo=F}
ma.key <- ma %>% 
  select(AUDIT_Total, meth_SDS, T_Ang_Total)

# Calculate mahalanobis distance
ma.key$mahal <- mahalanobis(ma.key, colMeans(ma.key), cov(ma.key))

# p-value for mahalanobis
ma.key$p <- pchisq(ma.key$mahal, df = 2, lower.tail = FALSE)

# tibble(
#   index = which(ma.key$p < .05),
#   mahal = ma.key$mahal[which(ma.key$p < .05)],
#   p = ma.key$p[which(ma.key$p < .05)]
# ) %>% kable(caption = "Mahalanobis < .05", align = "c") %>% kable_styling()
 
# Create a plotty plot
ma.key %>% 
  ggplot(aes(x = as.numeric(row.names(ma.key)), y = mahal, col = p < .05)) +
  geom_point() +
  geom_text(size = 2.5,
            label = as.numeric(row.names(ma)),
            nudge_y = 0.3,
            check_overlap = T) +
  scale_color_manual(values = c("black", "red")) +
  theme_light() +
  labs(x = "Index", y = "Mahalanobis Distance") +
  ggtitle("Mahalanobis Distance Plot")
```

##### Residual Leverage Plot {-}
```{r, echo = F}
ols_plot_resid_lev(m3.rm) +
  theme_light()
```

## Removed Outlier (Forced Entry)
**Omitting response 146**

```{r outlier.prep, include=F}
# Create column to easily remove rows
ma <- ma %>% 
  mutate(include = TRUE)
```

### Model Building
```{r buil.models.2}
ma.nout <- ma %>% mutate(include = ifelse(id == 146, FALSE, TRUE)) # Not including id 146

m0.2 <- lm(DDDI_Total~1, data = ma.nout[ma.nout$include == TRUE, ]) # Total SS
m1.2 <- lm(DDDI_Total~age + sex, data = ma.nout[ma.nout$include == TRUE, ]) # Model 1 (demographics)
m2.2 <- lm(DDDI_Total~age + sex + AUDIT_Total + meth_SDS, data = ma.nout[ma.nout$include == TRUE, ]) # Alcohol and MA use characteristics
m3.2 <- lm(DDDI_Total~age + sex + AUDIT_Total + meth_SDS + T_Ang_Total, data = ma.nout[ma.nout$include == TRUE, ]) # STAXI
```

### Sums of Squares
#### Total SS {-}
```{r ss.m0.2, echo=F}
anova(m0.2) %>% kable() %>% kable_styling()
```

#### Model SS {-}
```{r ss.models.2, echo=F}
anova(m1.2, m2.2, m3.2) %>% kable(align = "c") %>% kable_styling()
```
The model with Alcohol and SDS accounts for an additional 3088.06 Sum of Squares over and above the model including age and gender and this change was statistically significant (p < .01).  The $R^2$ increased by 24.6% in model 2.  

The model with Trait Anger accounts for an additional 1823.60 Sum of Squares controlling for demographics and alcohol/substance use and this too was statistically significant (p < .01).  The $R^2$ increased by 14.6%.  

### Model Summaries
#### Model 1 {-}
```{r model.sum.2, echo=F}
options(scipen = 999)
summary(m1.2)
```

#### Model 2 {-}
```{r model.2.2, echo=F}
summary(m2.2)
```

#### Model 3 {-}
```{r model.3.2, echo=F}
ols_regress(m3.2)
```

### Assumptions
#### Plot of Residuals {-}
```{r resid.plot.2, echo=F}
std.res.2 <- rstandard(m3.2)

m3.2.std.res <- tibble(
  Std.Res = std.res.2,
  Fitted = m3.2$fitted.values
)

m3.2.std.res %>% 
  ggplot(aes(x = Fitted, y = Std.Res)) +
  geom_point() +
  geom_hline(aes(yintercept = 0), linetype = "dashed", alpha = 0.5) +
  theme_light() +
  labs(x = "Fitted", y = "Standardised Residuals") +
  ggtitle("Plot of Residuals")

```

#### Colinearity (model 3) {-}
```{r}
tibble(Variable = c("Age", "Sex", "AUDIT-C", "SDS", "STAXI-T"),
       VIF = vif(m3.2))
```

#### Homogeneity {-}
```{r ass.homo.2, echo = F}
plot(m3.2, 1)
```

#### Linearity {-}

```{r ass.lin.2, echo=F}
plot(m3.2, 5)
```



#### Normality of Residuals {-}
```{r ass.norm.2, echo=F}
plot(m3.2, 2)
```

#### Outliers {-}
##### Residuals {-}
```{r ass.out.2, echo=F}
# Residuals by index
ma.nout[ma.nout$include == TRUE, ] %>% 
  ggplot(aes(x = as.numeric(row.names(ma.nout[ma.nout$include == TRUE, ])), y = std.res.2)) +
  geom_point() +
  geom_hline(aes(yintercept = 2), linetype = "dashed", alpha = 0.5) +
  geom_hline(aes(yintercept = -2), linetype = "dashed", alpha = 0.5) +
  geom_text(size = 3,
            label = as.numeric(row.names(ma.nout[ma.nout$include == TRUE,])),
            nudge_x = 0.25,
            nudge_y = 0.25,
            check_overlap = T) +
  theme_light() +
  labs(x = "Index", y = "Std Residuals") +
  ggtitle("Std. Residual by Index")
```

##### Cooks D {-}
```{r cooks.d.2, echo=F}
# Calculate Cook's D for each observation
cooksD <- cooks.distance(m3.2)

# Plot Cook's D
n <- nrow(ma.nout[ma.nout$include == TRUE,])

ma.nout[ma.nout$include == TRUE,] %>% 
  ggplot(aes(x = as.numeric(row.names(ma.nout[ma.nout$include == TRUE,])), y = cooksD)) +
  geom_point() +
  geom_hline(yintercept = 4/n, linetype = "dashed") +
  geom_text(size = 2.5,
            label = as.numeric(row.names(ma.nout[ma.nout$include == TRUE,])),
            nudge_y = 0.01,
            check_overlap = T) +
  theme_light() +
  labs(x = "Index", y = "Cooks Distance") +
  ggtitle("Cooks Distance")
```

##### Mahalanobis Distance {-}
```{r mahalanobis.2, echo=F}
ma.key.nout <- ma.nout[ma.nout$include == TRUE, ] %>% 
  select(age, sex, AUDIT_Total, meth_SDS, T_Ang_Total)

ma.key.nout <- ma.key.nout %>% 
  mutate(sex = ifelse(ma.key.nout$sex == "Male", 1, 0)) # Converting sex to numeric for calculations

# Calculate mahalanobis distance
ma.key.nout$mahal <- mahalanobis(ma.key.nout, colMeans(ma.key.nout), cov(ma.key.nout))


# p-value for mahalanobis
ma.key.nout$p <- pchisq(ma.key.nout$mahal, df = 4, lower.tail = FALSE)

tibble(
  index = which(ma.key.nout$p < .05),
  # id = which(ma.nout$id[which(ma.key.nout$p < .05),])
  mahal = ma.key.nout$mahal[which(ma.key.nout$p < .05)],
  p = ma.key.nout$p[which(ma.key.nout$p < .05)]
) %>% kable(caption = "Mahalanobis < .01", align = "c") %>% kable_styling()
 
# Create a plotty plot
ma.key.nout %>% 
  ggplot(aes(x = as.numeric(row.names(ma.key.nout)), y = mahal, col = p < .05)) +
  geom_point() +
  scale_color_manual(values = c("black", "red")) +
  theme_light() +
  labs(x = "Index", y = "Mahalanobis Distance") +
  ggtitle("Mahalanobis Distance Plot")
```

##### Residual Leverage Plot {-}
```{r resid.lev.plot.m3, echo=F}
ols_plot_resid_lev(m3.2) +
  theme_light()
```

## Removed Outlier (Reduced Model)
### Model Building
```{r nout.step}
ma.nout <- ma %>% filter(id != 146) # Not including id 146

m0.3 <- lm(DDDI_Total~1, data = ma.nout) # Total SS
m1.3 <- lm(DDDI_Total~age + sex, data = ma.nout) # Model 1 (demographics)
m2.3 <- lm(DDDI_Total~age + AUDIT_Total + meth_SDS, data = ma.nout) # Alcohol and MA use characteristics
m3.3 <- lm(DDDI_Total~AUDIT_Total + meth_SDS + T_Ang_Total, data = ma.nout) # STAXI
```

### Sums of Squares
#### Total SS {-}
```{r nout.step.ss, echo=F}
anova(m0.3) %>% kable(align = "c") %>% kable_styling()
```

#### Model SS {-}
```{r nout.step.ssm, echo=F}
anova(m1.3, m2.3, m3.3) %>% kable(align = "c") %>% kable_styling()
```
### Model Summaries
#### Model 1 {-}
```{r nout.sum.1, echo=F}
summary(m1.3)
```

#### Model 2{-}
```{r nout.sum.2, echo=F}
summary(m2.3)
```

#### Model 3 {-}
```{r nout.sum.3, echo=F}
ols_regress(m3.3)
```

### Assumptions
#### Plot of Residuals {-}
```{r nout.resid, echo=F}
std.res.3 <- rstandard(m3.3)

m3.3.std.res <- tibble(
  Std.Res = std.res.3,
  Fitted = m3.3$fitted.values
)

m3.3.std.res %>% 
  ggplot(aes(x = Fitted, y = Std.Res)) +
  geom_point() +
  geom_hline(aes(yintercept = 0), linetype = "dashed", alpha = 0.5) +
  theme_light() +
  labs(x = "Fitted", y = "Standardised Residuals") +
  ggtitle("Plot of Residuals")
```

#### Colinearity {-}
```{r nout.colin, echo=F}
tibble(
  Variable = c("AUDIT-C", "SDS", "STAXI-T"),
  VIF = vif(m3.3)
)
```

#### Homogeneity {-}
```{r nout.homo, echo=F}
plot(m3.3, 1)
```

#### Linearity {-}

```{r ass.lin.3, echo=F}
plot(m3.3, 5)
```

#### Normality of Residuals {-}
```{r ass.norm.3, echo=F}
plot(m3.3, 2)

shapiro.test(m3.3$residuals)
```

#### Outliers {-}
##### Residuals {-}
```{r ass.out.3, echo=F}
# Residuals by index
ma.nout %>% 
  ggplot(aes(x = as.numeric(row.names(ma.nout)), y = std.res.3)) +
  geom_point() +
  geom_hline(aes(yintercept = 2), linetype = "dashed", alpha = 0.5) +
  geom_hline(aes(yintercept = -2), linetype = "dashed", alpha = 0.5) +
  geom_text(size = 3,
            label = as.numeric(row.names(ma.nout)),
            nudge_x = 0.25,
            nudge_y = 0.25,
            check_overlap = T) +
  theme_light() +
  labs(x = "Index", y = "Std Residuals") +
  ggtitle("Std. Residual by Index")
```

##### Cooks D {-}
```{r cooks.d.3, echo=F}
# Calculate Cook's D for each observation
cooksD <- cooks.distance(m3.3)

# Plot Cook's D
n <- nrow(ma.nout)

ma.nout %>% 
  ggplot(aes(x = as.numeric(row.names(ma.nout)), y = cooksD)) +
  geom_point() +
  geom_hline(yintercept = 4/n, linetype = "dashed") +
  geom_text(size = 2.5,
            label = as.numeric(row.names(ma.nout)),
            nudge_y = 0.01,
            check_overlap = T) +
  theme_light() +
  labs(x = "Index", y = "Cooks Distance") +
  ggtitle("Cooks Distance")
```

## Robust Regression
Information on robust regression taken from <a href = https://stats.idre.ucla.edu/r/dae/robust-regression/> here. </a> and <a href = https://www.youtube.com/watch?v=qte9ASvgElI> here </a>

### Full model (Bisquare Weights)
```{r rlm.full, echo=F}
rbst.full <- robustbase::lmrob(DDDI_Total ~ age + sex + AUDIT_Total + meth_SDS + T_Ang_Total, data = ma, psi = "bisquare")
summary(rbst.full)
```

#### Standardised
```{r std.rbst, echo=F}
ma.std <- ma %>% 
  mutate(sex = ifelse(sex == "Male", 1, 0)) %>% 
  mutate(age_Std = scale(age),
         sex_Std = scale(sex),
         DDDI_Std = scale(DDDI_Total),
         AUDIT_Std = scale(AUDIT_Total),
         SDS_Std = scale(meth_SDS),
         STAXI_Std = scale(T_Ang_Total)
         )

rbst.std <- robustbase::lmrob(DDDI_Std ~ age_Std + sex_Std + AUDIT_Std + SDS_Std + STAXI_Std, data = ma.std)
summary(rbst.std)
rbst.std$coefficients
```


### Reduced Model 
```{r rlm.red, echo=F}
rbst.red <- robustbase::lmrob(DDDI_Total ~ AUDIT_Total + meth_SDS + T_Ang_Total, data = ma)
summary(rbst.red)
```

#### Standardised {-}
```{r std.robust.red, echo=F}
rbst.red.std <- robustbase::lmrob(DDDI_Std ~ AUDIT_Std + SDS_Std + STAXI_Std, data = ma.std)
summary(rbst.red.std)
rbst.red.std$coefficients
```

## Effect Sizes
### Full Model
Calculations for Cohen's $f^2$ were taken from <a href = https://www.spss-tutorials.com/effect-size/#linear-regression> here </a>.
```{r partial.cor, echo=F}
partial.vars <- ma %>% 
  mutate(sex = ifelse(sex == "Male", 1, 0)) %>% 
    select(DDDI_Total, age, sex, AUDIT_Total, meth_SDS, T_Ang_Total)

partial.cor.full <- ppcor::pcor(partial.vars)
part.full <- ppcor::spcor(partial.vars)

tibble(
  Variable = c("DDDI Total", "Age", "Sex", "AUDIT-C", "SDS", "STAXI-T"),
  Simple = data.frame(cor(partial.vars)) %>% pull(DDDI_Total),
  Partial = data.frame(partial.cor.full$estimate) %>% pull(DDDI_Total),
  Part = data.frame(part.full$estimate) %>% pull(DDDI_Total),
  f2 = (Part^2)/(1 - Part^2)
) %>% kable(align = "c") %>% kable_styling()
```

### Reduced Model
#### Model 1{-}
```{r m1.es, echo=F}
m1.rm.es <- ma %>%
  mutate(sex = ifelse(sex == "Male", 1, 0)) %>% 
  select(DDDI_Total, age, sex)

m1.rm.partial <- ppcor::pcor(m1.rm.es)
m1.rm.part <- ppcor::spcor(m1.rm.es)

tibble(
  Variable = c("DDDI Total", "Age", "Sex"),
  Simple = data.frame(cor(m1.rm.es)) %>% pull(DDDI_Total),
  Partial = data.frame(m1.rm.partial$estimate) %>% pull(DDDI_Total),
  Part = data.frame(m1.rm.part$estimate) %>% pull(DDDI_Total),
  f2 = (Part^2)/(1 - Part^2)
)
```

#### Model 2 {-}
```{r m2.es, echo=F}
m2.rm.es <- ma %>%
  select(DDDI_Total, age, AUDIT_Total, meth_SDS)

m2.rm.partial <- ppcor::pcor(m2.rm.es)
m2.rm.part <- ppcor::spcor(m2.rm.es)

tibble(
  Variable = c("DDDI Total", "Age", "AUDIT-C", "SDS"),
  Simple = data.frame(cor(m2.rm.es)) %>% pull(DDDI_Total),
  Partial = data.frame(m2.rm.partial$estimate) %>% pull(DDDI_Total),
  Part = data.frame(m2.rm.part$estimate) %>% pull(DDDI_Total),
  f2 = (Part^2)/(1 - Part^2)
)

```

#### Final Model {-}
```{r cors.reduced, echo=F}
reduced.vars <- ma %>% 
  select(DDDI_Total, AUDIT_Total, meth_SDS, T_Ang_Total)

partial.reduced <- ppcor::pcor(reduced.vars)
part.reduced <- ppcor::spcor(reduced.vars)

tibble(
  Variable = c("DDDI Total", "AUDIT-C", "SDS", "STAXI-T"),
  Simple = data.frame(cor(reduced.vars)) %>% pull(DDDI_Total),
  Partial = data.frame(partial.reduced$estimate) %>% pull(DDDI_Total),
  Part = data.frame(part.reduced$estimate) %>% pull(DDDI_Total),
  f2 = (Part^2)/(1 - Part^2)
) %>% kable(align = "c") %>% kable_styling()
```

### Full Model (Outlier Removed)
```{r effect.size.nout, echo=F}
full.nout <- ma.nout %>% 
  mutate(sex = ifelse(sex == "Male", 1, 0)) %>% 
  select(DDDI_Total, age, sex, AUDIT_Total, meth_SDS, T_Ang_Total)

partial.full.nout <- ppcor::pcor(full.nout)
part.full.nout <- ppcor::pcor(full.nout)

tibble(
  Variable = c("DDDI Total", "Age", "Sex", "AUDIT-C", "SDS", "STAXI-T"),
  Simple = data.frame(cor(full.nout)) %>% pull(DDDI_Total),
  Partial = data.frame(partial.full.nout$estimate) %>% pull(DDDI_Total),
  Part = data.frame(part.full.nout$estimate) %>% pull(DDDI_Total),
  f2 = (Part^2)/(1 - Part^2)
) %>% kable(align = "c") %>% kable_styling()
```

### Reduced Model (Outlier Removed)
```{r effect.size.nout.rem, echo=F}
red.nout <- ma.nout %>% 
  select(DDDI_Total, AUDIT_Total, meth_SDS, T_Ang_Total)

partial.red.nout <- ppcor::pcor(red.nout)
part.red.nout <- ppcor::pcor(red.nout)

tibble(
  Variable = c("DDDI Total", "AUDIT-C", "SDS", "STAXI-T"),
  Simple = data.frame(cor(red.nout)) %>% pull(DDDI_Total),
  Partial = data.frame(partial.red.nout$estimate) %>% pull(DDDI_Total),
  Part = data.frame(part.red.nout$estimate) %>% pull(DDDI_Total),
  f2 = (Part^2) / (1-Part^2)
) %>% kable(align = "c") %>% kable_styling()
```

## Comparing Model Fit (AIC)
```{r aic, echo=F}
rlm.full <- MASS::rlm(DDDI_Total ~ age + sex + AUDIT_Total + meth_SDS + T_Ang_Total, data = ma, method = "MM", psi = psi.bisquare)
rlm.red <- MASS::rlm(DDDI_Total ~ AUDIT_Total + meth_SDS + T_Ang_Total, data = ma, method = "MM", psi = psi.bisquare)

m3 <- lm(DDDI_Total~age + sex + AUDIT_Total + meth_SDS + T_Ang_Total, data = ma)

tibble(
  Model = c("Full", "Reduced", "Full (Outlier Removed)", "Reduced (Outlier Removed)", "Robust (Full)", "Robust (Reduced)"),
  AIC = c(AIC(m3), AIC(m3.rm), AIC(m3.2), AIC(m3.3), AIC(rlm.full), AIC(rlm.red))
) %>% kable(align = "c") %>% kable_styling()
```

## Best Subset of models
```{r best.subset, echo=F}
load("objects/m3.rda")
ols_step_best_subset(m3)
```

## Moderation TBC
```{r modertion, echo=F}
m0 <- lm(DDDI_Total~1, data = ma) # Total SS
m1 <- lm(DDDI_Total ~ age + sex, data = ma)
m2.rm <- lm(DDDI_Total ~ age + AUDIT_Total + meth_SDS, data = ma)
m3.rm <- lm(DDDI_Total ~ AUDIT_Total + meth_SDS + T_Ang_Total, data = ma)

# Centering Variables
ma.mod <- ma %>% 
  mutate(age.cen = age - mean(age),
         AUDIT.cen = AUDIT_Total - mean(AUDIT_Total),
         SDS.cen = meth_SDS - mean(meth_SDS),
         STAXI.cen = T_Ang_Total - mean(T_Ang_Total))

m2.rm.mod <- lm(DDDI_Total ~ AUDIT.cen + SDS.cen, data = ma.mod)
mint.rm <- lm(DDDI_Total ~ AUDIT.cen*meth_SDS, data = ma.mod)

anova(m2.rm.mod, mint.rm)

m3.rm.mod <- lm(DDDI_Total ~ AUDIT.cen + SDS.cen + STAXI.cen, data = ma.mod)
m4.rm.mod <- lm(DDDI_Total ~ AUDIT.cen + SDS.cen*STAXI.cen, data = ma.mod)
m5.rm.mod <- lm(DDDI_Total ~ SDS.cen + AUDIT.cen*STAXI.cen, data = ma.mod)
m6.rm.mod <- lm(DDDI_Total ~ STAXI.cen + AUDIT.cen*SDS.cen, data = ma.mod)

anova(m3.rm.mod, m4.rm.mod)
anova(m3.rm.mod, m5.rm.mod)
anova(m3.rm.mod, m6.rm.mod)

m3.mod <- lm(DDDI_Total ~ age.cen + sex + AUDIT.cen + SDS.cen + STAXI.cen, data = ma.mod)
m4.mod <- lm(DDDI_Total ~ sex + SDS.cen + STAXI.cen + age.cen*AUDIT.cen, data = ma.mod)
m5.mod <- lm(DDDI_Total ~ sex + AUDIT.cen + STAXI.cen + age.cen*SDS.cen, data = ma.mod)
m6.mod <- lm(DDDI_Total ~ sex + AUDIT.cen + SDS.cen + age.cen*STAXI.cen, data = ma.mod)
m7.mod <- lm(DDDI_Total ~ sex + AUDIT_Total + meth_SDS + age*T_Ang_Total, data = ma.mod)
m8.mod <- lm(DDDI_Total ~ age + meth_SDS + T_Ang_Total + sex*AUDIT_Total, data = ma)
m9.mod <- lm(DDDI_Total ~ age + T_Ang_Total + AUDIT_Total + sex*meth_SDS, data = ma)
m10.mod <- lm(DDDI_Total ~ age + meth_SDS + AUDIT_Total + sex*T_Ang_Total, data = ma)
m11.mod <- lm(DDDI_Total ~ age*sex + meth_SDS + AUDIT_Total + T_Ang_Total, data = ma)

anova(m3, m11.mod)
```
